+++ src/spinguin/_core/__init__.py
@@ -21,6 +21,7 @@ from ._hamiltonian import hamiltonian
 from ._liouvillian import liouvillian
 from ._molecule import Molecule
 from ._nmr_isotopes import (
+    add_isotope,
     atomic_mass,
     gamma,
     natural_abundance,
@@ -97,6 +98,7 @@ __all__ = [
     "Molecule",
 
     # nmr_isotopes
+    "add_isotope",
     "atomic_mass",
     "gamma",
     "natural_abundance",